+++ src/azure-cli/azure/cli/command_modules/appservice/_validators.py
@@ -119,3 +119,21 @@ def validate_add_vnet(cmd, namespace):
     if vnet_loc != webapp_loc:
         raise CLIError("The app and the vnet resources are in different locations. \
                         Cannot integrate a regional VNET to an app in a different region")
+
+
+def validate_asp_exists_in_rg(cmd, namespace):
+    import json
+    client = web_client_factory(cmd.cli_ctx)
+    serverfarm = namespace.name
+    resource_group_name = namespace.resource_group_name
+    asp = client.app_service_plans.get(resource_group_name, serverfarm, None, raw=True)
+    if asp.response.status_code != 200:
+        raise CLIError(asp.response.text)
+    # Isolated SKU is supported only for ASE
+    if namespace.sku in ['I1', 'I2', 'I3']:
+        # convert byte array to json
+        output_str = asp.response.content.decode('utf8')
+        res = json.loads(output_str)
+        if res.get('properties').get('hostingEnvironment') is None:
+            raise CLIError("The pricing tier 'Isolated' is not allowed in this resource group.Use this link to "
+                           "learn more: http://go.microsoft.com/fwlink/?LinkId=825764")