+++ src/azure-cli/azure/cli/command_modules/privatedns/commands.py
@@ -26,6 +26,11 @@ def load_command_table(self, _):
         client_factory=cf_privatedns_mgmt_record_sets
     )
 
+    network_privatedns_custom = CliCommandType(
+        operations_tmpl='azure.cli.command_modules.privatedns.custom#{}',
+        client_factory=cf_privatedns_mgmt_record_sets
+    )
+
     with self.command_group('network private-dns zone', network_privatedns_zone_sdk) as g:
         g.command('delete', 'delete', confirmation=True, supports_no_wait=True)
         g.show_command('show', 'get', table_transformer=transform_privatedns_zone_table_output)
@@ -39,7 +44,7 @@ def load_command_table(self, _):
         g.show_command('show', 'get', table_transformer=transform_privatedns_link_table_output)
         g.command('list', 'list', table_transformer=transform_privatedns_link_table_output)
         g.custom_command('create', 'create_privatedns_link', client_factory=cf_privatedns_mgmt_virtual_network_links, supports_no_wait=True)
-        g.generic_update_command('update', setter_name='update', custom_func_name='update_privatedns_link', supports_no_wait=True)
+        g.generic_update_command('update', setter_name='update_privatedns_link', setter_type=network_privatedns_custom, supports_no_wait=True)
         g.wait_command('wait')
 
     with self.command_group('network private-dns record-set') as g: