+++ src/azure-cli-core/HISTORY.rst
@@ -3,6 +3,16 @@
 Release History
 ===============
 
+2.0.75
+++++++
+* Add syntax check for --set parameter in generic update command.
+* Fix #10760: `az login` fails if subscription name is None
+* Minor change in DeploymentOutputLongRunningOperation
+
+2.0.74
+++++++
+* Introduce Debian Buster package
+
 2.0.73
 ++++++
 * Minor changes